requirements.txt
torch==2.0.0+cpu
torchvision==0.15.0+cpu
torchaudio==2.0.0+cpu
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
gradio>=4.0.0
Pillow>=9.0.0
sentencepiece>=0.1.99 